refactor: 이동 규칙 패키지 구조 변경 누락을 반영
